requirements.txt
streamlit 
moviepy 
edge-tts 
requests 
Pillow 
numpy 
watchdog 
